s25rttr: bump to EAPI 6, update cmake patch, drop games.eclass
Fixes: https://github.com/nE0sIghT/vortex-overlay/issues/15

+++ games-strategy/s25rttr/s25rttr-9999-r2.ebuild
@@ -2,8 +2,8 @@
 # Distributed under the terms of the GNU General Public License v2
 # $Header: /var/cvsroot/gentoo-x86/games-strategy/s25rttr/s25rttr-0.8.1.ebuild,v 1.1 2013/12/23 13:45:08 hasufell Exp $
 
-EAPI=5
-inherit eutils cmake-utils gnome2-utils games git-r3
+EAPI=6
+inherit eutils cmake-utils gnome2-utils git-r3
 
 DESCRIPTION="Open Source remake of The Settlers II game (needs original game files)"
 HOMEPAGE="http://www.siedler25.org/"
@@ -31,16 +31,13 @@ DEPEND="${RDEPEND}
 
 PATCHES=(
         "${FILESDIR}"/${P}-cmake.patch
-        "${FILESDIR}"/${P}-libutil.patch
-        "${FILESDIR}"/${P}-liblua.patch
-#        "${FILESDIR}"/${P}-boost.patch
 )
 
 src_prepare() {
 	# Ensure no bundled libraries are used
 	for file in $(ls ${S}/contrib/); do
-		# Preserve boost backports
-		if [ "${file}" != "backport" ]; then
+		# Preserve boost backports and kaguya
+		if [ "${file}" != "backport" -a "${file}" != "kaguya" ]; then
 			rm -r contrib/"${file}" || die
 		fi
 	done
@@ -63,16 +60,15 @@ src_configure() {
 	esac
 
 	local mycmakeargs=(
+		-DCMAKE_SKIP_RPATH=ON
+		-DRTTR_INSTALL_PREFIX=/usr
+		-DRTTR_DRIVERDIR="$(get_libdir)/${PN}"
+		-DRTTR_GAMEDIR="~/.${PN}/S2"
+		-DRTTR_LIBDIR="$(get_libdir)/${PN}"
 		-DCOMPILEFOR="linux"
 		-DCOMPILEARCH="${arch}"
 		-DCMAKE_SKIP_RPATH=YES
-		-DRTTR_PREFIX="${GAMES_PREFIX}"
-		-DRTTR_BINDIR="${GAMES_BINDIR}"
-		-DRTTR_DATADIR="${GAMES_DATADIR}"
-		-DRTTR_LIBDIR="$(games_get_libdir)/${PN}"
-		-DRTTR_DRIVERDIR="$(games_get_libdir)/${PN}"
-		-DRTTR_GAMEDIR="~/.${PN}/S2"
-		$(cmake-utils_use_build glfw GLFW_DRIVER)
+		-DBUILD_GLFW="$(usex glfw)"
 	)
 
 	cmake-utils_src_configure
@@ -83,39 +79,33 @@ src_compile() {
 	ln -s "${CMAKE_USE_DIR}"/RTTR "${CMAKE_BUILD_DIR}"/RTTR || die
 
 	cmake-utils_src_compile
-
-	mv "${CMAKE_USE_DIR}"/RTTR/{sound-convert,s-c_resample} "${T}"/ || die
 }
 
 src_install() {
 	cd "${CMAKE_BUILD_DIR}" || die
 
-	exeinto "$(games_get_libdir)"/${PN}
-	doexe "${T}"/{sound-convert,s-c_resample}
-	exeinto "$(games_get_libdir)"/${PN}/video
+	exeinto /usr/"$(get_libdir)"/${PN}
+	doexe s-c/src/sound-convert s-c/resample-1.8.1/src/s-c_resample
+	exeinto /usr/"$(get_libdir)"/${PN}/video
 	doexe driver/video/SDL/src/libvideoSDL.so
 	use glfw && doexe driver/video/GLFW/src/libvideoGLFW.so
-	exeinto "$(games_get_libdir)"/${PN}/audio
+	exeinto /usr/"$(get_libdir)"/${PN}/audio
 	doexe driver/audio/SDL/src/libaudioSDL.so
 
-	insinto "${GAMES_DATADIR}"
+	insinto /usr/share/"${PN}"
 	doins -r "${CMAKE_USE_DIR}"/RTTR
 
 	doicon -s 64 "${CMAKE_USE_DIR}"/debian/${PN}.png
-	dogamesbin src/s25client
-	make_desktop_entry "s25client" "Settlers RTTR" "${PN}"
+	dobin src/s25client
+	make_desktop_entry "s25client" "Settlers RTTR" "${PN}" "Game;StrategyGame" "Path=/usr/bin"
 	dodoc RTTR/texte/{keyboardlayout.txt,readme.txt}
-
-	prepgamesdirs
 }
 
 pkg_preinst() {
-	games_pkg_preinst
 	gnome2_icon_savelist
 }
 
 pkg_postinst() {
-	games_pkg_postinst
 	elog "Copy your Settlers2 game files into ~/.${PN}/S2"
 
 	gnome2_icon_cache_update
